# contracts/openapi/rhautt-nexus-v2.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Rhautt Nexus / 瑞合数智枢纽 API",
    "version": "1.0.0",
    "description": "Production v2 contract for 瑞诺瓦AI舒适家. Rhautt Comfort / 瑞合瑞德暖通科技集团 is the group expression, not the software name. 瑞诺瓦 is the C-end comfort-home system brand."
  },
  "servers": [
    {
      "url": "/",
      "description": "Same-origin production server"
    }
  ],
...
  "paths": {
...
    "/api/v2/quotation/generate": {
      "post": {
        "tags": [
          "Quote"
        ],
        "operationId": "generateQuotation",
        "summary": "生成报价（不持久化，纯计算）",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "design": {
                    "type": "object",
                    "additionalProperties": true
                  },
                  "devices": {
                    "type": "array",
                    "items": {
                      "type": "object",
                      "additionalProperties": true
                    }
                  },
                  "services": {
                    "type": "array",
                    "items": {
                      "type": "string"
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v2/auth/admin/users": {
      "get": {
        "tags": [
          "Auth"
        ],
        "operationId": "adminListUsers",
        "summary": "按当前租户范围查询账号",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "platform_admin",
          "hq_admin",
          "dealer_admin"
        ],
        "parameters": [
          {
            "name": "search",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "role",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "enum": [
                "active",
                "inactive",
                "suspended"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          }
        }
      },
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "adminCreateUser",
        "summary": "在当前租户范围创建账号",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "platform_admin",
          "hq_admin",
          "dealer_admin"
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonBody"
        },
        "responses": {
          "201": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "409": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
    "/api/v2/auth/admin/users/{id}/reset-password": {
      "post": {
        "tags": [
          "Auth"
        ],
        "operationId": "adminResetUserPassword",
        "summary": "在当前租户范围重置账号密码",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": [
          "platform_admin",
          "hq_admin",
          "dealer_admin"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "$ref": "#/components/requestBodies/JsonBody"
        },
        "responses": {
          "200": {
            "$ref": "#/components/responses/Success"
          },
          "400": {
            "$ref": "#/components/responses/Error"
          },
          "401": {
            "$ref": "#/components/responses/Error"
          },
          "403": {
            "$ref": "#/components/responses/Error"
          },
          "404": {
            "$ref": "#/components/responses/Error"
          }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "requestBodies": {
      "JsonBody": {
        "required": false,
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "additionalProperties": true
            }
          }
        }
      },
...
    },
    "responses": {
      "Success": {
        "description": "Successful 瑞诺瓦AI舒适家 envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/SuccessEnvelope"
            }
          }
        }
      },
...
      "Error": {
        "description": "Error 瑞诺瓦AI舒适家 envelope",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/ErrorEnvelope"
            }
          }
        }
      },
...
    },
    "schemas": {
...
      "SuccessEnvelope": {
        "type": "object",
        "required": [
          "success"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": true
          },
          "data": {}
        }
      },
...
      "ErrorEnvelope": {
        "type": "object",
        "required": [
          "success",
          "error"
        ],
        "properties": {
          "success": {
            "type": "boolean",
            "const": false
          },
          "error": {
            "type": "string"
          }
        }
      },
...
    }
  }
}